# identity/openapi/identity.yaml
openapi: 3.0.3
info:
  title: Identity API
  version: 0.1.0
  description: >
    Identity owns users, authentication, and the authenticated principal exposed
    to products. Future account-recovery and external-provider flows are
    tracked outside this active contract until their Keycloak-backed shape is
    designed.
  license:
    name: Proprietary
servers:
  - url: /api/v1
tags:
  - name: Identity
    description: Identity service metadata.
  - name: Users
    description: Identity-owned user records.
  - name: Sessions
    description: Login, current authenticated principal, and logout operations.
paths:
  /identity:
    get:
      tags: [Identity]
      operationId: getIdentityStatus
      summary: Get identity service status.
      responses:
        "200":
          description: Identity service status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceStatusResponse"
  /identity/users:
    post:
      tags: [Users]
      operationId: createUser
      summary: Create an identity user with password credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      tags: [Users]
      operationId: getUserByEmail
      summary: Get an identity user by email.
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /identity/users/provisional:
    post:
      tags: [Users]
      operationId: createProvisionalUser
      summary: Create an incomplete identity user for an invitation flow.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProvisionalUserRequest"
      responses:
        "201":
          description: Provisional user created or found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /identity/users/search:
    post:
      tags: [Users]
      operationId: searchUsers
      summary: Search identity users.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchUsersRequest"
      responses:
        "200":
          description: Users matching the search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /identity/users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: Get the current authenticated identity user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current identity user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Users]
      operationId: updateCurrentUser
      summary: Update current identity-owned profile fields.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCurrentUserRequest"
      responses:
        "200":
          description: User updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /identity/users/{userId}/completion:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [Users]
      operationId: completeProvisionalUser
      summary: Complete an invited identity user by adding profile data and a password.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteProvisionalUserRequest"
      responses:
        "201":
          description: Provisional user completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /identity/users/{userId}/provisional:
    parameters:
      - $ref: "#/components/parameters/UserId"
    delete:
      tags: [Users]
      operationId: cancelProvisionalUser
      summary: Cancel an incomplete invited identity user.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Provisional user cancelled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /identity/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      operationId: getUser
      summary: Get an identity user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Update identity-owned profile fields.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /identity/sessions:
    post:
      tags: [Sessions]
      operationId: authenticate
      summary: Sign in with email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticateRequest"
      responses:
        "201":
          description: Session created. Also sets an HTTP-only session cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedPrincipalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /identity/sessions/current:
    get:
      tags: [Sessions]
      operationId: getCurrentPrincipal
      summary: Get the current authenticated principal.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current authenticated principal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedPrincipalResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Sessions]
      operationId: deleteCurrentSession
      summary: Sign out the current session.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session deleted. Also clears the HTTP-only session cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      $ref: "../../common/openapi/errors.yaml#/components/responses/BadRequest"
    Unauthorized:
      $ref: "../../common/openapi/errors.yaml#/components/responses/Unauthorized"
    Forbidden:
      $ref: "../../common/openapi/errors.yaml#/components/responses/Forbidden"
    NotFound:
      $ref: "../../common/openapi/errors.yaml#/components/responses/NotFound"
    Conflict:
      $ref: "../../common/openapi/errors.yaml#/components/responses/Conflict"
  schemas:
    ServiceStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    UserStatus:
      type: string
      enum: [invited, active, disabled]
    UpdateUserStatus:
      type: string
      enum: [active, disabled]
    UserResponse:
      type: object
      required: [id, authorizationSubject, email, name, status, emailVerified, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        authorizationSubject:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
          format: uri
        status:
          $ref: "#/components/schemas/UserStatus"
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SearchUsersRequest:
      type: object
      properties:
        authorizationSubjects:
          type: array
          items:
            type: string
          uniqueItems: true
          default: []
    CreateUserRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
    CreateProvisionalUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    CompleteProvisionalUserRequest:
      type: object
      required: [name, password]
      properties:
        name:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        avatarUrl:
          type: string
          nullable: true
          format: uri
        status:
          $ref: "#/components/schemas/UpdateUserStatus"
    UpdateCurrentUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        avatarUrl:
          type: string
          nullable: true
          format: uri
    AuthenticateRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthenticationMethod:
      type: string
      enum: [password, oidc, saml]
    AuthenticatedPrincipalResponse:
      type: object
      required: [sessionId, user, grants, authenticationMethod, authProviderId, expiresAt]
      properties:
        sessionId:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/UserResponse"
        grants:
          type: array
          items:
            $ref: "#/components/schemas/ResourceGrantResponse"
        authenticationMethod:
          $ref: "#/components/schemas/AuthenticationMethod"
        authProviderId:
          type: string
          nullable: true
          format: uuid
        expiresAt:
          type: string
          nullable: true
          format: date-time
    ResourceGrantResponse:
      type: object
      required: [resource, actions]
      properties:
        resource:
          $ref: "#/components/schemas/GrantedResourceResponse"
        actions:
          type: array
          items:
            type: string
    GrantedResourceResponse:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
        id:
          type: string